feat(ui): dashboard and details routes no longer protected
By using the fee-sink as the sender address, simulate contract calls do note require the user to be connected.

Instead of showing a connect wallet menu on the root index route and redirecting to /dashboard, now the index route is the dashboard.
