+++ ui/src/components/AddValidatorForm.tsx
@@ -273,7 +273,7 @@ export function AddValidatorForm({ constraints }: AddValidatorFormProps) {
         },
       )
 
-      navigate({ to: '/dashboard' })
+      navigate({ to: '/' })
     } catch (error) {
       toast.error('Failed to create validator', { id: toastId })
       console.error(error)
